Minor Changes
Mudanças realizadas:

1° Conversão de Tsx para Jsx
2° Recriação dos componentes Header.jsx, footer.jsx, POButton.jsx, POMenssages.jsx.
3° Criação do componente SafeArea.jsx
