+++ src/Telas/Inicial.jsx
@@ -2,16 +2,16 @@ import { StyleSheet, Text, View, Image, TouchableOpacity } from 'react-native';
 
 {/*---------- <Componentes personalizados> ----------*/}
 import HeaderLogo from "../Components/Header";
-import POButtons from "../Components/POButtons";
-import {POTextLink} from "../Components/POText";
+import { POButton } from "../Components/POButton";
+import { POTextLink } from "../Components/POText";
 
 export default function App() {
 
   return (
     <View style={styles.container}>
       <HeaderLogo/>
-      <POButtons cor="orange" texto="Cadastre-se"/>
-      <POButtons cor="white" texto="Login" textColor="black"/>
+      <POButton styleBtn={'orange'} text={"Cadastre-se"}/>
+      <POButton styleBtn={'white'} text={"Login"}/>
       
       <POTextLink texto="É um vendedor?" link="Clique Aqui!"/>
     </View>
@@ -24,15 +24,5 @@ const styles = StyleSheet.create({
     alignItems: 'center',
     justifyContent: 'center',
   },
-
-  loginbuttons: {
-    height:50,
-    width:225,
-    backgroundColor:"orange",
-    borderRadius:30,
-    alignItems: "center",
-    justifyContent: "center",
-  },
-
 
 });
